Refactor firebase service
Split into:
- app.js (Initialisation and config)
- db.js (Firestore)
- auth.js (Authentication)
